backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy
Pillow
torch
torchvision
timm
